Move fully to MonsterFactory model
