+++ src/muncher/adventure/AdventureMunchHelpers.js
@@ -1,11 +1,11 @@
 import logger from "../../logger.js";
 import CompendiumHelper from "../../lib/CompendiumHelper.js";
 import PatreonHelper from "../../lib/PatreonHelper.js";
-import { parseCritters } from "../monsters.js";
 import { parseSpells } from "../spells.js";
 import { parseItems } from "../items.js";
 import AdventureMunch from "./AdventureMunch.js";
 import SETTINGS from "../../settings.js";
+import DDBMonsterFactory from "../DDBMonsterFactory.js";
 
 export default class AdventureMunchHelpers {
 
@@ -213,7 +213,8 @@ export default class AdventureMunchHelpers {
               if (tiers.all) {
                 logger.debug(`Importing missing ${type}s from DDB`, docIds);
                 AdventureMunch._progressNote(`Importing ${docIds.length} missing ${type}s from DDB`);
-                resolve(parseCritters(docIds));
+                const monsterFactory = new DDBMonsterFactory();
+                resolve(monsterFactory.processIntoCompendium(docIds));
               } else {
                 logger.warn(`Unable to import missing ${type}s from DDB - link to patreon or use your own proxy`, docIds);
                 ui.notifications.warn(`Unable to import missing ${type}s from DDB - link to patreon or use your own proxy`, { permanent: true });
